protect these controllers by error helper
  application, authorized application, client, user
add redirect_to_sign_in

+++ src/controllers/oauth/client_controller.cr
@@ -5,11 +5,13 @@ module OAuth
     end
 
     def index
+      return error if error?
       clients = OAuth::Client.all("WHERE user_id = $1", [current_user.id])
       render("src/views/oauth/client/index.slang")
     end
 
     def show
+      return error if error?
       if client = OAuth::Client.all("WHERE id = $1 and user_id = $2", [params["id"], current_user.id]).first
         render("src/views/oauth/client/show.slang")
       else
@@ -19,11 +21,13 @@ module OAuth
     end
 
     def new
+      return error if error?
       client = OAuth::Client.new
       render("src/views/oauth/client/new.slang")
     end
 
     def create
+      return error if error?
       client = OAuth::Client.new(params.to_h.select(["name", "redirect_uri", "scopes"]))
       client.user_id = current_user.id
 
@@ -37,6 +41,7 @@ module OAuth
     end
 
     def edit
+      return error if error?
       if client = OAuth::Client.all("WHERE id = $1 and user_id = $2", [params["id"], current_user.id]).first
         render("src/views/oauth/client/edit.slang")
       else
@@ -46,6 +51,7 @@ module OAuth
     end
 
     def update
+      return error if error?
       if client = OAuth::Client.all("WHERE id = $1 and user_id = $2", [params["id"], current_user.id]).first
         client.set_attributes(params.to_h.select(["name", "redirect_uri", "scopes"]))
         if client.valid? && client.save
@@ -62,6 +68,7 @@ module OAuth
     end
 
     def destroy
+      return error if error?
       if client = OAuth::Client.all("WHERE id = $1 and user_id = $2", [params["id"], current_user.id]).first
         client.destroy
       else
