+++ src/controllers/user_controller.cr
@@ -1,8 +1,8 @@
 class UserController < ApplicationController
   before_action do
     only [:edit, :update] do
-      unless authenticate_user! && current_user.id == params["id"].to_i
-        redirect_to root_path
+      unless current_user? && current_user.id == params["id"].to_i
+        redirect_to_sign_in
       end
     end
 
@@ -10,11 +10,13 @@ class UserController < ApplicationController
   end
 
   def index
+    return error if error?
     users = User.all
     render("index.slang")
   end
 
   def show
+    return error if error?
     if user = User.find params["id"]
       render("show.slang")
     else
@@ -24,11 +26,13 @@ class UserController < ApplicationController
   end
 
   def new
+    return error if error?
     user = User.new
     render("new.slang")
   end
 
   def create
+    return error if error?
     user = User.new(params.to_h.select(["email", "owner"]))
     user.password = params["password"]
 
@@ -42,6 +46,7 @@ class UserController < ApplicationController
   end
 
   def edit
+    return error if error?
     if user = User.find params["id"]
       render("edit.slang")
     else
@@ -51,6 +56,7 @@ class UserController < ApplicationController
   end
 
   def update
+    return error if error?
     if user = User.find(params["id"])
       old_password, user.password = user.password, params["password"]
       if old_password == params["old_password"] && params["password"] == params["confirm_password"] && user.valid? && user.save
@@ -67,6 +73,7 @@ class UserController < ApplicationController
   end
 
   def destroy
+    return error if error?
     if user = User.find params["id"]
       user.destroy
     else
